// giant-bpm-fe/src/types/flow.ts
export enum WorkflowNodeKey {
  Approval = "Approval",
  ParallelApproval = "ParallelApproval",
  Condition = "Condition",
  Subflow = "Subflow",
  Form = "Form",
  End = "End",
  Placeholder = "Placeholder",
  DummyNode = "DummyNode",
}
// ...
interface BaseNodeData extends Record<string, unknown> {
  key?: string;
  label?: string;
  description?: string;
  next: string | null;
  parents: string[] | null;
}
// ...
export interface SubflowNodeData extends BaseNodeData {
  workflowId: string | null;
  workflow: FlowDefinition | null;
}